lacon-diffusers/requirements.txt
diffusers>=0.21.0
transformers>=4.25.0
accelerate>=0.20.0
torch>=2.0.0
torchvision>=0.15.0
numpy
opencv-python
pillow
tqdm
einops
omegaconf
matplotlib
scipy